Move upload media to uploadFileScript and Change it to return promise and create appendFile to field for file field accepting multiple files
